requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sqlalchemy>=2.0.35
anthropic>=0.40.0
google-genai>=1.0.0
aiohttp>=3.10.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.10.0
bcrypt>=4.2.0
python-jose[cryptography]>=3.3.0
loguru>=0.7.0
aiosmtplib>=3.0.2
